LayoutTests/fast/dom/HTMLImageElement/image-natural-width-height-svg-expected.txt



PASS naturalWidth/Height of SVG in <img>, width/height in pixels
FAIL naturalWidth/Height of SVG in <img>, width in pixels; height unspecified assert_equals: expected "500x0" but got "300x150"
FAIL naturalWidth/Height of SVG in <img>, width in pixels; percentage height assert_equals: expected "500x0" but got "300x150"
PASS naturalWidth/Height of SVG in <img>, width/height in pixels; viewBox
FAIL naturalWidth/Height of SVG in <img>, width/height unspecified; viewBox assert_equals: expected "0x0" but got "800x600"
FAIL naturalWidth/Height of SVG in <img>, width in pixels; height unspecified; viewBox assert_equals: expected "400x0" but got "800x600"
